using less props by using function props
