+++ .Scribe/src/components/bookList/bookList.jsx
@@ -1,5 +1,4 @@
 import "./bookList.css";
-import { useState } from "react";
 import BookCard from "./bookCard/bookCard";
 
 const BookList = ({ books, className }) => {
