+++ .Scribe/src/components/header/header.jsx
@@ -2,27 +2,21 @@ import "./header.css";
 import DisplayBooks from "../displayBooks/displayBooks";
 
 const Header = ({
-  isSearchStarted,
   books,
-  setBooks,
   searchTerm,
   setisSearchStarted,
   setSearchTerm,
-  searchBooks,
+  doSearch,
 }) => {
   return (
     <header>
-      {isSearchStarted && books.length > 0 && (
-        <DisplayBooks
-          setisSearchStarted={setisSearchStarted}
-          books={books}
-          setBooks={setBooks}
-          searchTerm={searchTerm}
-          setSearchTerm={setSearchTerm}
-          isSearchStarted={isSearchStarted}
-          searchBooks={searchBooks}
-        />
-      )}
+      <DisplayBooks
+        setisSearchStarted={setisSearchStarted}
+        books={books}
+        searchTerm={searchTerm}
+        setSearchTerm={setSearchTerm}
+        doSearch={doSearch}
+      />
     </header>
   );
 };
